Update tomcat_dockerfile
In this code, the command "tar xvf apache-tomcat-10.0.27.tar.gz" is used instead of "tar xvfz apache*.tar.gz" and "mv apache-tomcat-10.0.27/* /opt/tomcat" is used instead of "mv apache-tomcat-10.0.2/* /opt/tomcat". The command "ls -l" is added after tar command to check whether the directory 'apache-tomcat-10.0.27' is created or not.

This should resolve the error message you were seeing, and the directory 'apache-tomcat-10.0.27' should be created and the files should be moved to the '/opt/tomcat/' directory.